Various changes
Remove BASEDIM_NONE since it is not required.
Fix typos.
Define some useful quantities.
A Quantity now defines an SI unit.
Add base() for a Quantity.
Add dimensionless for a Unit and a Quantity.
Add Reading and conversion between units.
Add parse()n for a string representation to a Reading.
Add UnitSystem as a collection of units.